Fixed trainer sprites, implemented ShinyDragonHunter's rework

+++ src/battle_controller_link_partner.c
@@ -261,7 +261,7 @@ static void LinkPartnerHandleHealthBarUpdate(u32 battler)
 static void LinkPartnerHandleIntroTrainerBallThrow(u32 battler)
 {
     u32 trainerPicId = GetLinkPlayerBackTrainerPicId(GetBattlerMultiplayerId(battler));
-    const u32 *trainerPal = gTrainerBackPicPaletteTable[trainerPicId];
+    const u16 *trainerPal = gTrainerBackPicPaletteTable[trainerPicId];
     // Link partner uses the same intro sequence as the player partner.
     BtlController_HandleIntroTrainerBallThrow(battler, 0xD6F9, trainerPal, 24, Controller_PlayerPartnerShowIntroHealthbox);
 }
